Update run_msf_addons.sh (#86)
Patch to `run_msf_addons.sh` to address issue Josh was facing when testing WF2 `Get Encounters` step. This step uses `http` adaptor because we query the OMRS FHIR api, which requires `baseUrl` in the auth configuration... but the `omrs` adaptor (used in the other workflow steps) uses `instanceUrl`. In the immediate term, Josh addressed this by manually adding a new credential with `baseUrl` to the WF2 in his deployed instance. 

Merge this change if you want to ensure deployment of this `project.yaml` will work instantly, without manual intervention. 

Otherwise next week (estimated date: Jan 22), there will be an `openmrs` adaptor release with an enhancement that we can leverage and apply to the `project.yaml` file so that _all_ OMRS workflow steps can use the `openmrs` adaptor. If no rush, then you can wait until this upcoming change, and not merge this PR.